Initial Defines Setup, Housekeeping
Set up a header file to contain toggleable defines that should wrap
around modded code.

NQMod uses a GUID generated using Visual Studio that is different from
the original game's GUID, as required by Firaxis for all DLL mods.

Added all of the filters from the VS2008 solution to the VS2013
solution.

Added new functions that enable and control minidump generation.

Added the parsing of double-type variables from the game database,
seeing as the engine supports it, but Firaxis decided to not add it into
the DLL for some reason.

Added a few more mathematical constants that can be called via macro
instead of having to be recalculated using cmath functions (like e).