First turn fix, same turn wonder fix, barbarians cannot move immediately after spawning, human cities default to production focus
This commit contains 4 changes:

NQM_RANDOM_FIRST_TURN randomises who gets first turn. At the beginning
of each turn, human players are given a random ordering. Based on their
position in this ordering, they can not use units until a certain delay
after the turn has begun. For example, the 1st player in the ordering
can move their units immediately, but the 3rd player must wait 1.5
seconds. In this way, the effects of connection stability on first move
are made insignificant. Instead, whoever has the shorter delay has first
move, and this is randomly allocated, which is much fairer.

NQM_BARBARIANS_MOVE_BEFORE_SPAWNING is a simple fix to prevent
barbarians from spawning and moving in the same turn. This prevents
things like barbarians spawning and then immediately taking a worker,
before a player is able to react.

NQM_SAME_TURN_WORLD_WONDERS_DECIDED_BY_PRODUCTION_OVERFLOW makes it so
that if more than one player finishes the same world wonder on the same
turn, the player with the highest excess production is the one who gets
it. Tie breakers are decided randomly. This is much fairer than
previously, where for example the host would always get the wonder.

NQM_HUMAN_CITIES_PRODUCTION_FOCUS_DEFAULT is a tiny convenience change
that sets cities to production focus by default when they are settled by
human players. This is an action that is taken by all players anyway, so
this is just to save them some time.

These changes have been listed in the new _Defines.h, which makes it
easy to switch them on and off (credit to Delnar_Ersike for this
method).