Minor Fixes and QoL Tweaks
Fixed a badly setup conditional in CvPlot::canSeePlot().

Switched DangerPlots to use an array instead of an FFastVector to both
improve performance and cause the game to not crash when loading a
version of the DLL compiled with no optimizations.

Added a dedicated configuration to compile a version of the DLL without
most optimizations, letting the developer see all memory values during
debugging.