Tweaked Governor's Cost Tiebreaker, Yield Weighing
Optimized fruitstrike's original tiebreaker for the city governor that
was based on tiles' purchasing costs to no longer require the tile list
to be looped through and some of its members erased.

Yield weighing for the city governor no longer takes city focus into
account because there is no way to avoid a particular yield focus have
an indirect effect on other yields without inflating the value of yields
to the city governor.