Proper cached yields, More Accurate low-yield RAs
Research Agreements that have a low beaker amount invested will no
longer have an artificially low beaker reward.

Yields are no longer processed at the end of players' turns. Instead,
they are again processed at the beginning, but influence, tourism, and
religious spread are all processed at the beginning of players' turns as
well instead of at the beginning of the game turn for everyone.

Deals are no longer processed for everyone at once, they are instead
processed on a per-player basis at the beginning of the player's turn
after yields have been cached. With the exception of Research
Agreements, all deals are processed in two stages when they expire, one
for each player. When the deal expires, only the items sent to the
player whose current turn it is will be removed, and when the other
player's turn is processed, the other items in the deal will be removed.
For example, if a Defensive Pact between Player A and Player C is
expiring, Player C will still have a defensive pact from Player A during
Player A's turn while Player A will not have a defensive pact from
Player C; Player A's defensive pact to Player C will be removed at the
start of Player C's turn. Research Agreements are resolved
simultaneously, but they feed into the players' overflow beakers instead
of being applied directly to tech progress. Wartime declarations follow
an odd behavior: if the game is in simultaneous mode, deals are
cancelled like normal, but if the game is not in simultaneous mode, only
items to the player(s) who declared war are removed during the
declaration; items from the aggressor player(s) to the defending
player(s) are only removed at the start of the defending player(s) turn,
so that the aggressor(s) cannot abuse the player-based deal processing
system.