Tweaks to City Governor
Osmosis effects on plot influence score now only apply from plots that
are revealed to the player, so players can no longer tell that there
might be a resource tile in nearby fog based on where the city governor
wants to expand.

The fact that city focus influences yield scores is now controlled by a
define that can be toggled on or off easily.