Better Hybrid fixes, Constification
Constified getNumGameTurnActive() and countNumHumanGameTurnActive().

Introduced a function that returns true if there are no players that are
marked active. This function runs faster than getting the number of
total active players and checking if that number is 0.

Optimized a few more functions related to Better Hybrid.

Removed the player-activating function for Better Hybrid from the loop
that iterates through the current turn order and instead made the part
of updateMoves() that activates simultaneous mode players be responsible
for player activation in Better Hybrid.