City, CityCitizens Fixes
Citizen manager self-consistency check now has a threshold of 1 instead
of 0.

Citizen manager will now account for Guruship and similar beliefs. If
the city's primary religion has a belief that gives it extra yield if
the city has a specialist, that amount of extra yield is added to the
value of the first specialist.

Fixed the bug where breaking railroad connections would not remove the
city's production bonus from railroads.