Relative Turn Timers
Add an in-game option that makes custom turn timers relative when
ticked. This means that instead of setting the timer of all turns to the
custom amount, the default turn timer value is multiplied by the custom
amount and divided by 100. This effectively allows players to make turn
times longer or shorter, while still letting the turn time scale up as
the game progresses: for example, a setting of 50 would halve the turn
timer, a setting of 80 will multiply turn times by 0.8, a setting of 150
will multiply turn times by 1.5, etc.