Added Auto-Pause on Active Disconnect Option
Added an in-game option that should automatically pause the game if an
active, alive, non-observer, human player is disconnected from a
non-sequential, Pitboss game. If the player is kicked and there are no
other disconnected players, then the paused game should resume. If two
or more players disconnect, everyone needs to be reconnected or kicked
before the game will resume.