Replaced LFSR Attempt with Better LC Constants
Civ5's engine is extremely fickle when it tests that the game's RNG
"works", i.e. it will not load in games if the get() method differs too
much from its internal standard. It really didn't like any attempts at
LFSRs or non-linear LCs, so the attempts were scrapped.

However, the LC constants were set to a value that would need a modulus
of 2^31, which wasn't present in the original code. Introducing a
modulus step made the game's engine not like the RNG, so instead the
multiplication and offset constants have been changed to values that
would need a modulus of 2^32, which is automatically achieved if the
seed is stored as an unsigned int.