Fixed Possible Crash on Exit + Optimizations
Fixed a possible crash on game exit and performed optimizations in a
single move by replacing the Active Leagues FStaticVector with a single
Active League object. Deallocation of this object is more reliable
(fixing the crash) and accessing it is faster than going through an
FStaticVector to access the only possible active league anyway (the
optimization part).