Replaced Lua with LuaJIT
Replaced the game's Lua 5.1.4 hooks and DLL with LuaJIT 2.0.4, an
implementation of Lua 5.1.5 that uses a JIT compiler to speed code up.
Seeing as Lua code is the second big performance bottleneck in Civ5
(after the pathfinder) and that crashes often occur due to issues in the
Lua code, this should hopefully improve stability and performance.
LuaJIT 2.0.4 has been stable since May 2015.