Couple Reversions
Reverted to the older version of clearing all A* nodes of cached
information, as the newer version wasn't actually doing what it was
meant to (address issues of certain nodes not being cleared when
civilians were attempted to be moved).

Re-enabled the POD switch for one of the three vectors that originally
had it in the first change, since the data corruption bug only happened
when the other two were tagged as POD (since they both contain members
that have non-trivial constructors).