Added Safety Measure to Fix Unworkable City Tile Bug
There was originally a bug in the game where if you acquired a city
through a peace deal, the city's tile would be unworkable unless the
city was razed. This was fixed in another change, but just in case that
change does not work, a safety measure was added that should definitely
make sure the city's tile remains workable.