Fixed Specialist Control, Reduced Population Target
Fixed an oversight that was causing manual specialist control to not
work.

Reduced the population that a city wants to reach before it lowers its
minimum food amount from 5 pop to 2 pop. This means that only cities
with a single point of population will always want at least 2 excess
food.