On DoW change for Trade Units and Turn Timers
When war is declared against a team, trade units are only destroyed with
their trade routes if the war was manually declared by the trade units'
owners. If the trade route owner is the DoW recipient or if the DoW
happened as a result of a Defensive Pact, the trade units are returned
to their respective owners.

Added a feature where if war is declared in simultaneous mode with less
than 60 seconds on the turn timer, the turn timer is refilled to 60
seconds. This feature's define is disabled, but can be enabled in the
future.