Sorted Out Citizen Manager Issues
Fixed a bug that was causing all specialists (including unemployed ones)
to have a baseline value of 8 for the citizen manager.

If a city is set to food production, it will still assign a minimal
value to non-excess food yield, even if it would contribute no
production toward the project.

Sorted out food yield modifiers and effective food yield from reduced
consumption when scoring specialist slots to be representative of the
values the player would actually experience.

Fixed an issue with the citizen manager that was messing up the scoring
of gold yield converted to (higher scoring) science yield when the
player's negative gold per turn would cause negative science.

Fixed an issue that was causing the citizen manager to incorrectly
assign happiness score to an eventh specialist slot when checking for
the worst scoring, occupied specialist slot with half specialist
unhappiness enabled. E.g., because of rounding issues, if there are 4
occupied specialist slots in a city owned by a player with half
specialist unhappiness, removing one specialist won't increase
unhappiness because the half specialist unhappiness effect rounds
specialist population up, but this wasn't accounted for previously when
scoring specialist slots.