Citizens Auto-Reallocate on Food Production Change
When a city switches on or off of food production mode (i.e. it starts
producing a unit that diverts excess food to production when it didn't
before or stops producing these types of units), the citizen manager
instantly reallocates its citizens. If cached yields are enabled, this
is purely a quality-of-life change, since the city would reallocate
citizens at the start of the turn anyway, but now the player will
instantly see this reallocation on their end before they end their turn.