Intercept Withdraw Tweaked
Intercept withdraw changed so that the target health is an easily
modifiable amount instead of hard-set to 50 and that it accounts for air
repair as well: if the interceptor is out of interceptions after the air
sweep happens, the interceptor has air repair, and the interceptor's
owner's turn is still active (meaning the interceptor would heal before
its next interception), the interceptor's heal rate is added to its HP
when checking for whether or not it should withdraw. Additionally, the
target was increased to 55 health (from 50), which corresponds with the
HP required for a fighter to get one-shot exactly 50% of the time from
an opposing air sweep conducting by a fighter with equal base strength
and full health.