Citizen Manager Improvements, Fixes, Expansions
Improved the accuracy of the scoring mechanism behind calculating the
effects of extra happiness on non-happiness yields.

Improved the scoring of happiness yield by having it work off of the
finite integral of the desired happiness scoring function instead of
relying on a single-value version of that function.

When a city is set to food production, initial citizen reallocation
happens assuming that the city is not on food production, and food
production is only considered during the self-consistency check after
the initial citizen reallocation. This should greatly improve the
ability to get ideal yields during food production.

When a player's yields or happiness changes, their cities automatically
perform a self-consistency check, which looks like their citizens are
reallocated (but runs significantly faster than a full reallocation).