Stability Improvement, New AI Gimp
Added a new optional AI gimp that causes the AI to always accept white
peace offers from human players.

Forced classes that are strictly subservient to CvCity (i.e. any
instance of those classes is always attached to an instance of CvCity)
to assign pointers to their parent CvCity class in their constructor
instead of just in their init() function (which need not always be
called).