Puppet Logic Improvement, Optimizations
Slightly optimized the code for the AI selecting the initial weights for
buildings to build.

Replaced the code that gave a 2x weight multiplier to maintenance-free
buildings for puppets with code that increases or decreases a building's
weight via a logistic function based on the effective maintenance of the
building (i.e. building maintenance - extra gold the building would
generate).

Puppets are now able to build undesirable, non-wonder buildings (e.g.
Barracks) if there's nothing else they want to build.