Misc. Minor Fixes
Reformatted various lines of code to be contained in fewer lines. Added
a few minor fixes, such as Guruship scoring when a specialist is to be
removed, minor performance improvements for CS and barbarian turn
processing, and minor performance improvements for citizen slot scoring,
patched up a rare Radaring case.