Fixes to Citizen Yield Scoring on Slot Removal
When a currently worked slot's food yield is checked, it will now
properly recognize what the city's total food yield value would be
without the slot. That is, instead of calculating how much a duplicate
of that worked slot would be worth at the moment, the game will
calculate how much the city would lose by unworking the slot. Similar
fixes were implemented for gold yield when it is to be converted to
science and bonus happiness.