Religion Fixes and Optimizations
Optimized a few pieces of Religion code.

Added extra checkers to the Lua code that manually adds religious
beliefs so that it properly makes sure that beliefs can be added. This
avoids potential sync issues when using mods that rely on these Lua
functions instead of the network functions tied to adding beliefs.