Counterblockade Added
Added a new twist to blockades. If a would-be blockaded tile is closer
to a valid allied unit (i.e. naval unit in the same area) than the
closest valid blockading unit, the tile is not blockaded. This allows
ships to counterblockade to a certain extent, but it's limited enough
that a single allied ship cannot counter a large blockade on its own.