Air Sweep, Counterattack, City Yield, cont'd Fixes
Moved unit promotion testing to the function that changes XP so now it
is always tested when XP is changed instead of needing a separate call
each time.

Relocated city yield change effects from CvPlot to CvCity so that it
isn't rounded down to the nearest yield each time, i.e. it allows for
hundredths-level yield changes. This also means city yield changes don't
show up in the city tile yield.

Ranged counterattacks made by cities now cannot happen if the city would
not be able to attack due to resistance, being razed, or having 0 HP.

Added an in-game option for using the binomial RNG for combat rolls,
compensated for by a 4x increase in the range of random results
possible. The RNG first generates a random number between 0 and 59 using
a binomial distribution, then multiplies the result by 80x and adds a
random number between 0 and 79 (via flat RNG) as noise. This keeps rolls
down to a manageable 61 while still retaining a binomial-like
distribution. The number is re-zeroed to make sure the average result
would still have a net zero effect on damage even after the 4x range
increase.

Fixed a bug that was causing workers to build at 2x speed whenever they
started building something new. This messes up both EUI and the regular
UI, and unfortunately there's no way to fix displays for both at the
same time, so instead, efforts have been made to minimize the effect on
both. For EUI, the initial build turn count is 1 less than it actually
is, i.e. when a worker is to build something new, the displayed turn
amount required is 1 less than it actually is. For the regular UI,
intermittent turn amounts are incorrect, i.e. turn counts are incorrect
once a worker starts building an improvement (the incorrectness varies
on how much work the worker has done), but initial turn counts are all
correct, i.e. the amount of turns displayed when the worker is to build
something is the actual amount of turns required.

Melee ships can now properly counterattack when eligible against ranged
city attacks or attacks originating from units in enemy cities.

Altered the counterattack behavior between city attackers and melee
defenders: instead of forcing the melee defender to slam into the city,
the melee counterattack simply happens with a multiplicative -60%
penalty, which is the same penalty that cities have when they perform
ranged attacks.

Ranged attacks between a city and a unit no longer cause two messages to
pop up.

Air sweeps against ground interceptors will now properly apply damage to
the interceptor, display combat results, and give XP for the encounter
to the ground interceptor.

Fixed melee units not being able to attack cities.

When an existing citadel would be acquired from a culture bomb, then if
the existing citadel has any tiles surrounding it that are owned by the
citadel's current owner, would not be acquired from the same culture
bomb, and do not have citadels on them, then the existing citadel will
not be flipped by the culture bomb.

The UI will now display a city's ranged strength when they are to attack
a target instead of the city's melee strength.