WLtKD Caching, Faith Auto-Purchasing Fix
We Love the King Days can now trigger properly in a lot more different
cases. For example, if a lux is awarded after completing a tech,
completing a world wonder, completing a building in another city, or as
a reward for a spy action, WLtKDs will no longer trigger with a 1-turn
delay.

Faith auto-purchasing now happens as the last thing in a player's turn
processing, i.e. it's done even after replay stats are recorded. This
makes sure that faith auto-purchasing is never strictly better than
manually purchasing with faith, so it will always strictly be a QoL
feature.