Few NQMP Improvements
The "allow extra religions" option is replaced with an "all players can
found religions" option, which should also play nice with observer
players. The option no longer has a hard cap of 8, it now allows as many
religions to be founded as there are filled, non-observer slots in a
game.

Shifted around some of the WIP code for a UA where creating certain
types of units grants influence to all CSs a player has met, so that 1)
it runs faster, and 2) it hopefully doesn't crash when it activates.