Modified Binomial Combat RNG, Updated Defines
The binomial combat RNG option will now be completely powered by the
binomial RNG, rather than use a hybrid of both binomial and flat RNGs.
In addition, the 4x range multiplier has been replaced with a variable
range multiplier so that the range set by the game for flat distribution
represents a range of ±3σ under binomial distribution.

In layman's terms, this means combat rolls can now technically result in
any number from -1576 to 1624 (from 24 to 36), but players are 68.3%
likely to get a roll between 28 and 32, 95.4% likely to get a roll
between 26 and 34, 99.7% likely to get a roll between 24 and 36, 99.99%
likely to get a roll between 22 and 38, etc. By contrast, under flat
distribution, players have a 33.3% chance to get a roll between 28 and
32 and a 66.7% chance to get a roll between 26 and 34.