Cached Yields Altered for Stability
The int arrays initialized to the heap and used for storing cached
yields at the city level have been replaced with three, "vanilla"
integers. Private integer members are much easier to handle than int
arrays, especially because they're stored on the stack instead of on the
heap.

When cached yields are enabled, the default size for DealList is
increased to 128 elements (from 20) to compensate for the fourth new
type of global DealList and to avoid allocating new memory on-the-fly as
much as possible.