Heavy Charge Downhill Portability
Improved the portability of the code responsible for allowing units to
heavy charge when attacking downhill. The bulk of the code is now
wrapped inside CvUnit::IsCanHeavyCharge(pFromPlot, pToPlot) in a way
that if either pFromPlot or pToPlot are NULL, the function is identical
to CvUnit::IsCanHeavyCharge(). This also makes it so that giving
situational heavy charge to units will be easier in the future: only
IsCanHeavyCharge(pFromPlot, pToPlot) needs to be updated, since
everything else relies on that one function (yay for portable code!).