Defines Update, Re-enabling, Culling
Re-enabled all previously disabled defines that cannot possibly cause
engine crashes: their changes do not add anything that would be stored
on the heap (even temporarily), and they do not add, remove, or change
any object member. In case such a define was tied to a toggleable
option, that option was also re-added.

Separated out citadel resistance to culture bombs and disabled it with
"//" instead of inside a comment block. This is to indicate that the
define should be stable, but it isn't necessarily desirable.

Culled defines that do not correspond to any code.