Fixed Barbarian Camps Spawning Close to Players
Fixed a bug that allowed barbarian camps spawned on the first turn to
spawn closer to players than intended. Barbarian camps aren't supposed
to spawn within 4 tiles of a player's capitol, but this did not extend
to starting settlers when the world is populated with barbarian camps on
the barbarians' first turn.